♻️ [Refactoring]: #455 Phase 11 decode_escape 単体テストを tests/decode_escape.rs に分離
純関数 decode_escape の単体テスト 16 件を
literal_string/tests/decode_escape.rs に切り出し、tests.rs ハブに
`mod decode_escape;` 宣言を追加。pub(super) シンボルは
`use super::super::*;` 経由で descendant 可視性により参照可能。
